requirements.txt
asgiref==3.10.0
Django==5.2.7
djangorestframework==3.16.1
python-decouple==3.8
sqlparse==0.5.3
